A far more robust scheme for dealing with multi-disc sets.
Musicbrainz now has an advanced relation type that indicates the previous
and next discs in a multi-disc set. Instead of using the clunky and
unreliable method of looking up all releases with a matching ASIN (which
lead to issues when ASINs were ambiguous) we now use the PartOfSet relation
type and walk the list of releases.

This also should avoid the bug where the target directory was being created
but the rename failed as it couldn't get the number of discs.