flanamer: Be smarter about building the list of files to rename.
This removes the "hard-coded" filename pattern that flacnamer used to use. It
now simply lists all files in the source directory with matching extensions (at
the moment this is just all files ending in .flac). It assumes that tracks
are listed in sort-order. We also now do some basic checking on number of files
to rename versus number of files in the musicbrainz release.

This is the start of turning flacnamer into a more generic tool.