flacnamer: Remove the requirement for a source TOC.
This allows flacnamer to rename a directory even if there is no TOC,
i.e. renaming an album that not been ripped using ripcd/2flac. If a
TOC exists in the source directory it is used just like before. If there
is no TOC, then albumidentify is used to guess the release.