Make sure we've loaded track filenames
If you force a release with --release-id, then the tracks aren't actually
loaded into the disc structure by the time we decide we're going to process
them.  This change says "if we've not loaded the tracks now, and we need them
for the next step, force loading of them now using a similar system to
how albumidentify does it.