Change the way we deal with the "album" tag.
Instead of setting the album name to the full disc string, e.g.
"Stadium Arcadium (disc 1: Jupiter)" we now set ALBUM to just
the album name, use DISC and DISCNUMBER tags for the disc number
and add a DISCNAME tag for the name of the disc, if applicable.

While there doesn't seem to be support out there for a DISCNAME tag
it does seem that setting the ALBUM tag to the full string is wrong,
as all of the discs in an album should be able to be collated under
one title.