Tighten checks on all tracks, deal with OoO tracks.
All tracks now pass through a much stronger "verify_track" that
tries to make sure that all the tracks are valid before considering them.

Also, if we get a series of tracks out of order, see if we can find another
release, possibly with them in order.  If we get one with them in order
then give that as a success, if we get another one with them out of order
then don't accept either of them as being possible.